+++ changelog.md
@@ -4,6 +4,7 @@
 * Moved chicken food items to item tag `iguanatweaksreborn:chicken_food_items`
   * Also fixed not being able to use any item on chickens
 * Increased experience from experience bottles
+* Moved HardCrops feature to Crops
 
 ## 3.5.2
 * Players are no longer able to pillar **only if** there are monsters nearby